refactor: move shared metadata helpers from dbstats to utils and update imports
